simplify storage key types
